100% ConfigModule implementation❤️

+++ Runtime/ConfigModule/ConfigManagers/LocalDynamicConfigManager.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
@@ -29,6 +30,8 @@ namespace LSCore.ConfigModule
 
     public class LocalDynamicConfigManager<T> : BaseConfigManager<T>, ILocalConfigManager where T : LocalDynamicConfig, new()
     {
+        protected override string Tag => $"{base.Tag}.{Path.GetFileNameWithoutExtension(GetFullFileName(fullPath))}";
+
         protected virtual string GetPath(string path)
         {
             return path;
@@ -42,6 +45,7 @@ namespace LSCore.ConfigModule
         protected string fullPath;
         
         private bool wasLoaded;
+        private JObject token;
 
         public T Config
         {
@@ -59,11 +63,23 @@ namespace LSCore.ConfigModule
             return $"{path}.json";
         }
 
+        private string FullFileNameMeta => $"Full file name: {GetFullFileName(fullPath)}";
         public virtual void Load()
         {
+            SetMeta(FullFileNameMeta);
+            Log("Loading");
+            
+            if (cached == null)
+            {
+                Log("Config created");
+                cached = new T();
+                cached.AddMigrations();
+            }
+
             string json;
             
-            (string current, string target) data;
+            //TODO: Implement cool and convenient logic of path migration
+            /*(string current, string target) data;
             if (!Migrator.Path.TryGet(fullPath, out data))
             {
                 var fullFileName = GetFullFileName(fullPath);
@@ -73,7 +89,6 @@ namespace LSCore.ConfigModule
                 }
                 else
                 {
-                    cached ??= new T();
                     cached.SetDefault();
                     wasLoaded = true;
                     return;
@@ -82,25 +97,52 @@ namespace LSCore.ConfigModule
             else
             {
                 var currentPath = GetFullFileName(data.current);
-                json = File.ReadAllText(currentPath);
-                File.Delete(currentPath);
                 var targetPath = GetFullFileName(data.target);
-                File.WriteAllText(targetPath,json);
-            }
 
-            var token = JToken.Parse(json);
-            Migrator.Type<T>.Migrate(token);
-            Migrator.Migrate(fullPath, token);
-            json = token.ToString(Formatting.None);
-            
-            Deserialize(json);
+                if (File.Exists(targetPath))
+                {
+                    throw new Exception("File is already exists");
+                }
+                
+                if (File.Exists(currentPath))
+                {
+                    json = File.ReadAllText(currentPath);
+                    File.Delete(currentPath);
+                }
+                else
+                {
+                    cached.SetDefault();
+                    wasLoaded = true;
+                    return;
+                }
+                
+                Save(targetPath, json);
+            }*/
+
+            json = File.ReadAllText(GetFullFileName(fullPath));
+            SetMeta($"{FullFileNameMeta}\nJson:\n{json}");
+            Log("Read json");
+            token = JObject.Parse(json);
+            var wasMigrated = Migrator.Type<T>.Migrate(token);
+            wasMigrated |= Migrator.Migrate(fullPath, token);
+            if (wasMigrated)
+            {
+                SetMeta($"{FullFileNameMeta}\nJson:\n{token}");
+                Log("Migrated");
+            }
+            Deserialize(token);
             wasLoaded = true;
         }
 
         public virtual void Save()
         {
-            var json = Serialize();
+            var json = Serialize(token);
             string fullFileName = GetFullFileName(fullPath);
+            Save(fullFileName, json);
+        }
+
+        protected void Save(string fullFileName, string json)
+        {
             Directory.CreateDirectory(Path.GetDirectoryName(fullFileName)!);
             File.WriteAllText(fullFileName,json); 
         }
@@ -117,6 +159,17 @@ namespace LSCore.ConfigModule
         {
             string fullFileName = GetFullFileName(fullPath);
             File.Delete(fullFileName);
+        }
+        
+        
+        /*public void AddPathMigration(string newPath)
+        {
+            Migrator.Path.Add(fullPath, GetPath(newPath));
+        }*/
+        
+        public void AddMigration(Action<JToken> migrator)
+        {
+            Migrator.Add(fullPath, migrator);
         }
     }
 }
